Move error boundary screen to views and remove unecessary library
