+++ package.json
@@ -141,7 +141,6 @@
     "react-native-device-info": "3.1.4",
     "react-native-elevated-view": "0.0.6",
     "react-native-emoji": "1.3.1",
-    "react-native-error-boundary": "^1.1.3",
     "react-native-fade-in-image": "1.4.1",
     "react-native-flash-message": "0.1.11",
     "react-native-fs": "^2.16.6",
